+++ .github/workflows/test.yml
@@ -33,22 +33,6 @@ jobs:
 
           make install-shellcheck
 
-      - name: Cache libsecp256k1
-        id: cache-libsecp256k1
-        uses: actions/cache@v3
-        with:
-          path: /usr/local/lib/libsecp256k1*
-          key: secp256k1-linux-v1
-
-      - name: Install libsecp256k1
-        if: steps.cache-libsecp256k1.outputs.cache-hit != 'true'
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y build-essential autoconf libtool git
-          cd /tmp && git clone --depth 1 https://github.com/bitcoin-core/secp256k1.git
-          cd secp256k1 && ./autogen.sh && ./configure --enable-module-recovery
-          make -j$(nproc) && sudo make install && sudo ldconfig
-
       - name: Install golangci-lint
         run: |
           curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh \
@@ -97,21 +81,6 @@ jobs:
           go install github.com/fzipp/gocyclo/cmd/gocyclo@latest
           brew install glfw
         shell: bash
-
-      - name: Cache libsecp256k1
-        id: cache-libsecp256k1
-        uses: actions/cache@v3
-        with:
-          path: /usr/local/lib/libsecp256k1*
-          key: secp256k1-macos-v1
-
-      - name: Install libsecp256k1
-        if: steps.cache-libsecp256k1.outputs.cache-hit != 'true'
-        run: |
-          brew install autoconf automake libtool
-          cd /tmp && git clone --depth 1 https://github.com/bitcoin-core/secp256k1.git
-          cd secp256k1 && ./autogen.sh && ./configure --enable-module-recovery
-          make -j$(sysctl -n hw.ncpu) && sudo make install
 
       - name: Install golangci-lint
         run: |
